initial world for Yoku's Island Adventure
This can generate a save file for a Yoku's Island Adventure run which
runs correctly and appears to match the spoiler.

Logic is pulled from the built-in randomizer and so _should_ be correct,
but unit tests have not been written yet, and the region names are based
on the game's internal IDs.

The multiworld functionality is still a work in progress.